changed: put HandlerContext in separate compile unit
