+++ opm/input/eclipse/Schedule/Schedule.hpp
@@ -43,8 +43,6 @@
 #include <opm/input/eclipse/Schedule/WriteRestartFileEvents.hpp>
 #include <opm/input/eclipse/Units/UnitSystem.hpp>
 
-#include "src/opm/input/eclipse/Schedule/MSW/WelSegsSet.hpp"
-
 namespace Opm
 {
     namespace Action {
@@ -63,6 +61,7 @@ namespace Opm
     class GTNode;
     class GuideRateConfig;
     class GuideRateModel;
+    class HandlerContext;
     enum class InputErrorAction;
     class ParseContext;
     class Python;
@@ -491,71 +490,6 @@ namespace Opm
         void dump_deck(std::ostream& os) const;
 
     private:
-        struct HandlerContext {
-
-            const ScheduleBlock& block;
-            const DeckKeyword& keyword;
-            const std::size_t currentStep;
-            const std::vector<std::string>& matching_wells;
-            const bool actionx_mode;
-            const ParseContext& parseContext;
-            ErrorGuard& errors;
-            SimulatorUpdate* sim_update{nullptr};
-            const std::unordered_map<std::string, double>* target_wellpi{nullptr};
-            std::unordered_map<std::string, double>* wpimult_global_factor{nullptr};
-            WelSegsSet* welsegs_wells{nullptr};
-            std::set<std::string>* compsegs_wells{nullptr};
-            const ScheduleGrid& grid;
-
-            /// \param welsegs_wells All wells with a WELSEGS entry for checks.
-            /// \param compegs_wells All wells with a COMPSEGS entry for checks.
-            HandlerContext(const ScheduleBlock& block_,
-                           const DeckKeyword& keyword_,
-                           const ScheduleGrid& grid_,
-                           const std::size_t currentStep_,
-                           const std::vector<std::string>& matching_wells_,
-                           bool actionx_mode_,
-                           const ParseContext& parseContext_,
-                           ErrorGuard& errors_,
-                           SimulatorUpdate* sim_update_,
-                           const std::unordered_map<std::string, double>* target_wellpi_,
-                           std::unordered_map<std::string, double>* wpimult_global_factor_,
-                           WelSegsSet* welsegs_wells_,
-                           std::set<std::string>* compsegs_wells_)
-            : block(block_)
-            , keyword(keyword_)
-            , currentStep(currentStep_)
-            , matching_wells(matching_wells_)
-            , actionx_mode(actionx_mode_)
-            , parseContext(parseContext_)
-            , errors(errors_)
-            , sim_update(sim_update_)
-            , target_wellpi(target_wellpi_)
-            , wpimult_global_factor(wpimult_global_factor_)
-            , welsegs_wells(welsegs_wells_)
-            , compsegs_wells(compsegs_wells_)
-            , grid(grid_)
-            {}
-
-            void affected_well(const std::string& well_name);
-            void record_well_structure_change();
-
-            /// \brief Mark that the well occured in a  WELSEGS keyword
-            void welsegs_handled(const std::string& well_name)
-            {
-                if (welsegs_wells)
-                    welsegs_wells->insert(well_name, keyword.location());
-            }
-
-            /// \brief Mark that the well occured in a  COMPSEGS keyword
-            void compsegs_handled(const std::string& well_name)
-            {
-                if (compsegs_wells)
-                    compsegs_wells->insert(well_name);
-            }
-
-        };
-
         // Please update the member functions
         //   - operator==(const Schedule&) const
         //   - serializationTestObject()
