+++ src/opm/input/eclipse/Schedule/KeywordHandlers.cpp
@@ -96,6 +96,7 @@
 #include <opm/input/eclipse/Parser/ParserKeywords/V.hpp>
 #include <opm/input/eclipse/Parser/ParserKeywords/W.hpp>
 
+#include "HandlerContext.hpp"
 #include "Well/injection.hpp"
 
 #include <algorithm>
@@ -150,7 +151,7 @@ namespace {
         throw OpmInputError("AQUFETP is not supported as SCHEDULE keyword", handlerContext.keyword.location());
     }
 
-    void Schedule::handleAQUFLUX(Schedule::HandlerContext& handlerContext) {
+    void Schedule::handleAQUFLUX(HandlerContext& handlerContext) {
         // auto& aqufluxs = this->snapshots.back().aqufluxs;
         auto& aqufluxs = this->snapshots.back().aqufluxs;
         for (const auto& record : handlerContext.keyword) {
@@ -159,7 +160,7 @@ namespace {
         }
     }
 
-    void Schedule::handleBCProp(Schedule::HandlerContext& handlerContext) {
+    void Schedule::handleBCProp(HandlerContext& handlerContext) {
         auto& bcprop = this->snapshots.back().bcprop;
         for (const auto& record : handlerContext.keyword) {
             bcprop.updateBCProp(record);
@@ -2122,7 +2123,7 @@ Well{0} entered with 'FIELD' parent group:
         }
     }
 
-    void Schedule::handleWINJMULT(Opm::Schedule::HandlerContext& handlerContext) {
+    void Schedule::handleWINJMULT(HandlerContext& handlerContext) {
         for (const auto& record : handlerContext.keyword) {
             const std::string& wellNamePattern = record.getItem("WELL_NAME").getTrimmedString(0);
             const auto well_names = wellNames(wellNamePattern);
